Update UI for integrations setup (#1052)
* Stub catalog reader interface


* Add basic catalog functionality to new catalog reader


* Refactor validation logic with a deeper interface


* Refactor validation logic with a deeper interface


* Remove redundant test.

This test is unneeded after 12c4bcf


* Add tests for new validators


* Make better failure mode for invalid objects


* Generalize Result type


* Convert backend to use catalog reader (unstable)


* Repair tests for integrations class (unstable)


* Refactor repository for new integration interface


* Fix outer repository and backend tests


* Add tests for sample data


* Add CatalogReader JavaDocs


* Repair integrations builder


* Remove extra commented test


* Remove unnecessary log statement


* Repair getSchemas behavior to return correct type

Let it be known at on this day, with this commit,
I have truly grokked why we don't use `any` in typescript.


* Add tests for getSchemas


* Add tests for asset and sample data backend methods


* Break flyout validation methods out of constructing method


* Add tests for extracted flyout methods


* Switch validation method to use ValidationResult


* Swap out flyout for hello-world setup page


* Add basic step incrementing


* Add basic field skeleton for each step


* Add a cancel button


* Add config type to developing form


* Flatten integration config


* Add sample data table modal


* Add toggle for standard and advanced asset config


* Simplify imports


* Refactor major class names


* (WIP) begin refactoring functionality into adaptor


* Finish migrating functionality to data adaptor


* Rename integration types for more clarity


* Refactor component usage


* Connect forms to config state


* Fix filetype selector


* Remove hardcoded name in path


* Write one snapshot test


* Add more tests


* Fix test naming


* Update obsolete snapshots


* Move integration creation helpers to own file


* Break out integration creation methods


* Isolate more create_integration helpers


* Simplify setup form


* Add data source picker items


* Add better selector logic


* Add queries for data sources


* Switch from selector to combobox


* Update snapshots


* Connect validation button to data source validation method


* Reimplement add integration button


* Temporarily remove validate button


* Simplify dynamic table term selection


* Remove unused validate code


* Undo wildcard import


* Switch from proxy to dataconnections endpoint


* Remove unused table fields


* Switch dataconnections base to const


* Add console proxy to route constants


* Update snapshots


* Move color to constants


* Move index name validation to constants and improve matching


* Move test constants to test constants


---------
