+++ public/components/integrations/home.tsx
@@ -3,17 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useState } from 'react';
+import React from 'react';
 import { HashRouter, Route, RouteComponentProps, Switch } from 'react-router-dom';
-import { EuiGlobalToastList } from '@elastic/eui';
-import { Toast } from '@elastic/eui/src/components/toast/global_toast_list';
-import { Integration } from './components/integration';
 import { TraceAnalyticsCoreDeps } from '../trace_analytics/home';
 import { ChromeBreadcrumb } from '../../../../../src/core/public';
 import { AvailableIntegrationOverviewPage } from './components/available_integration_overview_page';
 import { AddedIntegrationOverviewPage } from './components/added_integration_overview_page';
 import { AddedIntegration } from './components/added_integration';
-import { SetupIntegrationStepsPage } from './components/setup_integration';
+import { SetupIntegrationPage } from './components/setup_integration';
+import { Integration } from './components/integration';
 
 export type AppAnalyticsCoreDeps = TraceAnalyticsCoreDeps;
 
@@ -63,7 +61,18 @@ export const Home = (props: HomeProps) => {
               />
             )}
           />
-          <Route exact path={'/available/:id/setup'} render={() => <SetupIntegrationStepsPage />} />
+          <Route
+            exact
+            path={'/available/:id/setup'}
+            render={(routerProps) => (
+              <SetupIntegrationPage
+                integration={{
+                  name: routerProps.match.params.id,
+                  type: 'logs', // TODO
+                }}
+              />
+            )}
+          />
         </Switch>
       </HashRouter>
     </div>
